update to cmake build
