+++ Sync/CMAKE_README.txt
@@ -1,0 +1,21 @@
+To generate build files for your platform run in a command terminal:
+cd Build
+cmake -G GENERATOR ..
+
+Valid generators are:
+Windows:
+"Visual Studio 12 2013 Win64"
+"Visual Studio 14 2015 Win64"
+"Visual Studio 15 2017 Win64"
+Mac:
+"Xcode"
+Linux: (see note below)
+"Unix Makefiles"
+
+Example: cmake -G "Visual Studio 12 2013 Win64" ..
+
+For Linux Only:
+On linux, Debug and Release options are generated by cmake and must be specified like so:
+cmake -G "Unix Makefiles" -DCMAKE_BUILD_TYPE=Release ..
+or
+cmake -G "Unix Makefiles" -DCMAKE_BUILD_TYPE=Debug ..
